make separate data loading utility for 2.0 json->numpy conversion
