+++ tests/test_map_api_v2.py
@@ -12,7 +12,6 @@ from argoverse.map_representation.map_api_v2 import (
     ArgoverseMapV2,
     PedCrossing,
     VectorLaneSegment,
-    point_arr_from_points_list_dict,
 )
 
 TEST_DATA_ROOT = Path(__file__).parent.resolve() / "test_data"
@@ -219,43 +218,3 @@ class TestArgoverseMapV2(unittest.TestCase):
         # fmt: on
         np.testing.assert_allclose(vector_da[:4], expected_first4_vertices)
 
-
-def test_point_arr_from_points_list_dict() -> None:
-    """Ensure dictionary with lane boundary polyline parameterization is lifted to Numpy array correctly."""
-    lane_segment = {
-        "right_lane_boundary": {
-            "points": [
-                {"x": 919.19, "y": 150.0, "z": -21.42},
-                {"x": 916.42, "y": 171.79, "z": -21.67},
-            ]
-        }
-    }
-    right_ln_bound = point_arr_from_points_list_dict(lane_segment["right_lane_boundary"]["points"])
-    # fmt: off
-    expected_right_ln_bound = np.array(
-        [
-            [919.19, 150.0, -21.42],
-            [916.42, 171.79, -21.67]
-        ]
-    )
-    # fmt: on
-    np.testing.assert_allclose(right_ln_bound, expected_right_ln_bound)
-
-
-def test_point_arr_from_points_list_dict() -> None:
-    """Ensure dictionary with lane boundary polyline parameterization is lifted to Numpy array correctly."""
-    # fmt: off
-    points_list_dict = [
-        {"x": 5504, "y": 3291, "z": -6},
-        {"x": 5491, "y": 3263, "z": -5}
-    ]
-    # fmt: on
-    arr = point_arr_from_points_list_dict(points_list_dict)
-    # fmt: off
-    gt_arr = np.array(
-        [
-            [5504, 3291, -6],
-            [5491, 3263, -5]
-        ])
-    # fmt: on
-    np.testing.assert_allclose(arr, gt_arr)
